Adds new project template selection
Implements a dialog for selecting a project template or sample.

This change introduces a dropdown menu in the "New Project" dialog,
allowing users to choose from a list of predefined templates and
samples. The selected template's data is then used to initialize
the new project, offering a more streamlined and user-friendly
experience when starting a new project.